# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  github_com_AVGsync_keyfreestay_backend_internal_model.HousingImage:
    properties:
      HousingID:
        description: UUID объекта жилья.
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      ID:
        description: UUID изображения.
        example: 880e8400-e29b-41d4-a716-446655440000
        type: string
      ImageURL:
        description: Публичный URL изображения, сформированный из storage key.
        example: https://cdn.keyfreestay.local/housing/550e8400-e29b-41d4-a716-446655440000/7a45d9c2-0000-4000-9000-123456789abc.jpg
        type: string
      Position:
        description: Позиция изображения в галерее.
        example: 0
        type: integer
      StorageKey:
        description: Storage key файла в S3/MinIO.
        example: /housing/550e8400-e29b-41d4-a716-446655440000/7a45d9c2-0000-4000-9000-123456789abc.jpg
        type: string
    type: object
  github_com_AVGsync_keyfreestay_backend_internal_model.HousingResponse:
    properties:
      address:
        description: Адрес объекта.
        example: Москва, ул. Тверская, 12
        type: string
      amenities:
        description: Список удобств объекта.
        example:
        - wifi
        - smart_lock
        items:
          enum:
          - kitchen
          - wifi
          - air_conditioner
          - parking
          - washing_machine
          - heating
          - smart_lock
          - smart_lighting
          - smart_thermostat
          - voice_assistant
          type: string
        type: array
      cancellation_policy:
        description: Правила отмены бронирования.
        enum:
        - flexible
        - moderate
        - strict
        example: flexible
        type: string
      created_at:
        description: Дата создания объекта.
        example: "2026-05-15T12:00:00Z"
        type: string
      description:
        description: Подробное описание жилья.
        example: Светлая квартира рядом с метро, бесконтактное заселение, быстрый
          Wi-Fi.
        type: string
      housing_type:
        description: Тип объекта жилья.
        enum:
        - apartment
        - house
        - office
        example: apartment
        type: string
      id:
        description: UUID объекта жилья.
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      images:
        description: Изображения объекта жилья.
        items:
          $ref: '#/definitions/github_com_AVGsync_keyfreestay_backend_internal_model.HousingImage'
        type: array
      latitude:
        description: Широта объекта.
        example: 55.7558
        type: number
      longitude:
        description: Долгота объекта.
        example: 37.6173
        type: number
      max_guests:
        description: Максимальное количество гостей.
        example: 4
        type: integer
      price_per_night:
        description: Цена за одну ночь.
        example: 6500
        type: number
      rating_avg:
        description: Средняя оценка объекта.
        example: 4.8
        type: number
      rating_count:
        description: Количество оценок.
        example: 12
        type: integer
      status:
        description: Статус объекта.
        enum:
        - draft
        - published
        - unpublished
        example: published
        type: string
      title:
        description: Название объекта.
        example: Современная квартира в центре
        type: string
      updated_at:
        description: Дата последнего обновления объекта.
        example: "2026-05-15T12:30:00Z"
        type: string
      user_id:
        description: UUID владельца объекта.
        example: 440e8400-e29b-41d4-a716-446655440000
        type: string
    type: object
...
  github_com_AVGsync_keyfreestay_backend_internal_model_request.CreateHousingRequest:
    properties:
      address:
        description: Адрес объекта жилья в свободной форме.
        example: Москва, ул. Тверская, 12
        maxLength: 500
        minLength: 3
        type: string
      description:
        description: Подробное описание жилья, правил и особенностей проживания.
        example: Светлая квартира рядом с метро, бесконтактное заселение, быстрый
          Wi-Fi.
        maxLength: 5000
        type: string
      housing_type:
        description: Тип объекта жилья. Если не передан, БД использует значение по
          умолчанию `apartment`.
        enum:
        - apartment
        - house
        - office
        example: apartment
        type: string
      title:
        description: Название объекта для каталога и карточки жилья.
        example: Современная квартира в центре
        maxLength: 200
        minLength: 3
        type: string
    type: object
...
  github_com_AVGsync_keyfreestay_backend_internal_model_request.UpdateHousingRequest:
    properties:
      address:
        description: Адрес объекта жилья в свободной форме.
        example: Москва, ул. Тверская, 12
        maxLength: 500
        minLength: 3
        type: string
      amenities:
        description: Список удобств объекта.
        example:
        - wifi
        - smart_lock
        items:
          enum:
          - kitchen
          - wifi
          - air_conditioner
          - parking
          - washing_machine
          - heating
          - smart_lock
          - smart_lighting
          - smart_thermostat
          - voice_assistant
          type: string
        type: array
      cancellation_policy:
        description: Правила отмены бронирования.
        enum:
        - flexible
        - moderate
        - strict
        example: flexible
        type: string
      description:
        description: Подробное описание жилья, правил и особенностей проживания.
        example: Светлая квартира рядом с метро, бесконтактное заселение, быстрый
          Wi-Fi.
        maxLength: 5000
        type: string
      housing_type:
        description: Тип объекта жилья.
        enum:
        - apartment
        - house
        - office
        example: apartment
        type: string
      latitude:
        description: Широта объекта для карты.
        example: 55.7558
        type: number
      longitude:
        description: Долгота объекта для карты.
        example: 37.6173
        type: number
      max_guests:
        description: Максимальное количество гостей.
        example: 4
        maximum: 50
        type: integer
      price_per_night:
        description: Цена за одну ночь в валюте сервиса.
        example: 6500
        type: number
      status:
        description: Статус объекта. `published` доступен только когда заполнены обязательные
          поля для публикации.
        enum:
        - draft
        - published
        - unpublished
        example: published
        type: string
      title:
        description: Название объекта для каталога и карточки жилья.
        example: Современная квартира в центре
        maxLength: 200
        minLength: 3
        type: string
    type: object
...
  github_com_AVGsync_keyfreestay_backend_internal_model_response.HousingListItem:
    properties:
      address:
        description: Адрес объекта.
        example: Москва, ул. Тверская, 12
        type: string
      housing_type:
        description: Тип объекта жилья.
        enum:
        - apartment
        - house
        - office
        example: apartment
        type: string
      id:
        description: UUID объекта жилья.
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      max_guests:
        description: Максимальное количество гостей.
        example: 4
        type: integer
      price_per_night:
        description: Цена за одну ночь.
        example: 6500
        type: number
      rating_avg:
        description: Средняя оценка объекта.
        example: 4.8
        type: number
      rating_count:
        description: Количество оценок.
        example: 12
        type: integer
      status:
        description: Статус объекта.
        enum:
        - draft
        - published
        - unpublished
        example: published
        type: string
      thumbnail_url:
        description: Публичный URL первого изображения объекта, если изображение есть.
        example: https://cdn.keyfreestay.local/housing/550e8400-e29b-41d4-a716-446655440000/photo.jpg
        type: string
      title:
        description: Название объекта.
        example: Современная квартира в центре
        type: string
    type: object
  github_com_AVGsync_keyfreestay_backend_internal_model_response.HousingListResponse:
    properties:
      items:
        description: Краткие карточки объектов жилья.
        items:
          $ref: '#/definitions/github_com_AVGsync_keyfreestay_backend_internal_model_response.HousingListItem'
        type: array
      total:
        description: Общее количество объектов в ответе.
        example: 42
        type: integer
    type: object
...
info:
  contact:
    email: support@keyfreestay.local
    name: KeyFreeStay API Support
  description: |-
    REST API сервиса KeyFreeStay для краткосрочной аренды жилья без ключей.

    API покрывает регистрацию и вход, профиль пользователя, объекты жилья, изображения жилья, бронирования, платёжные методы и публичную контактную заявку.
    Схема авторизации: вызовите `POST /auth/login`; API установит JWT в HttpOnly cookie `token`. Защищённые маршруты читают JWT из cookie.
    Swagger UI доступен только при `DEBUG=true`: `/swagger`, JSON спецификация — `/swagger/doc.json`.
  title: KeyFreeStay API
  version: "1.0"
paths:
...
  /housing:
    delete:
      description: |-
        Удаляет объект жилья текущего пользователя по query-параметру `id`.
        Удалять объект может только владелец. После удаления сервис также удаляет связанные изображения из S3/MinIO.
      parameters:
      - description: UUID объекта жилья
        example: 550e8400-e29b-41d4-a716-446655440000
        in: query
        name: id
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Пустой ответ со статусом 200
          schema:
            type: string
        "400":
          description: missing housing ID — id не передан
          schema:
            type: string
        "401":
          description: unauthorized — cookie `token` отсутствует, просрочена или недействительна
          schema:
            type: string
        "500":
          description: failed to delete housing — объект не найден, не принадлежит
            пользователю или не удалён
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Удалить объект жилья
      tags:
      - Жильё
    get:
      description: |-
        Возвращает список объектов жилья, доступный текущему пользователю.
        Пользователь с ролью `user` получает только опубликованные объекты. Владелец/администратор получает свои объекты.
        В списке возвращается краткая карточка и публичный URL первого изображения, если изображение есть.
      produces:
      - application/json
      responses:
        "200":
          description: Список объектов жилья
          schema:
            $ref: '#/definitions/github_com_AVGsync_keyfreestay_backend_internal_model_response.HousingListResponse'
        "401":
          description: unauthorized — cookie `token` отсутствует, просрочена или недействительна
          schema:
            type: string
        "500":
          description: failed to get housing list — список не получен
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Получить список объектов жилья
      tags:
      - Жильё
    patch:
      consumes:
      - application/json
      description: |-
        Частично обновляет объект жилья текущего пользователя.
        ID объекта передаётся в query-параметре `id`. Изменять объект может только его владелец.
        Для публикации (`status=published`) объект должен иметь обязательные поля: название, адрес, координаты, цену за ночь и максимум гостей.
      parameters:
      - description: UUID объекта жилья
        example: 550e8400-e29b-41d4-a716-446655440000
        in: query
        name: id
        required: true
        type: string
      - description: Поля объекта жилья, которые нужно изменить
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AVGsync_keyfreestay_backend_internal_model_request.UpdateHousingRequest'
      produces:
      - text/plain
      responses:
        "200":
          description: Пустой ответ со статусом 200
          schema:
            type: string
        "400":
          description: missing housing ID / invalid request body — не передан id или
            тело запроса некорректно
          schema:
            type: string
        "401":
          description: unauthorized — cookie `token` отсутствует, просрочена или недействительна
          schema:
            type: string
        "500":
          description: failed to update housing — объект не найден, не принадлежит
            пользователю или не прошёл ограничения БД
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Обновить объект жилья
      tags:
      - Жильё
    post:
      consumes:
      - application/json
      description: |-
        Создаёт черновик объекта жилья для текущего пользователя.
        При создании можно передать тип жилья, название, описание и адрес. Остальные поля заполняются через PATCH /housing.
        Новый объект принадлежит пользователю из cookie `token`.
      parameters:
      - description: Начальные данные объекта жилья
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_AVGsync_keyfreestay_backend_internal_model_request.CreateHousingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 'ID созданного объекта: {\"id\":\"550e8400-e29b-41d4-a716-446655440000\"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: invalid request body — тело запроса не JSON или структура не
            совпадает со схемой
          schema:
            type: string
        "401":
          description: unauthorized — cookie `token` отсутствует, просрочена или недействительна
          schema:
            type: string
        "500":
          description: failed to create housing — объект не создан
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Создать объект жилья
      tags:
      - Жильё
  /housing/{id}:
    get:
      description: |-
        Возвращает полную карточку объекта жилья с параметрами, рейтингом и списком изображений.
        Для каждого изображения сервис формирует публичный URL из storage key.
      parameters:
      - description: UUID объекта жилья
        example: 550e8400-e29b-41d4-a716-446655440000
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Полная карточка объекта жилья
          schema:
            $ref: '#/definitions/github_com_AVGsync_keyfreestay_backend_internal_model.HousingResponse'
        "400":
          description: missing housing ID — id не передан в path
          schema:
            type: string
        "401":
          description: unauthorized — cookie `token` отсутствует, просрочена или недействительна
          schema:
            type: string
        "500":
          description: failed to get housing — объект не найден или произошла внутренняя
            ошибка
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Получить объект жилья по ID
      tags:
      - Жильё
  /housing/{id}/images:
    delete:
      description: |-
        Удаляет изображение из S3/MinIO и запись о нём в БД.
        ID объекта передаётся в path, storage key изображения — в query-параметре `key`.
        Удалять изображение может только владелец объекта жилья.
      parameters:
      - description: UUID объекта жилья
        example: 550e8400-e29b-41d4-a716-446655440000
        in: path
        name: id
        required: true
        type: string
      - description: Storage key изображения в S3/MinIO
        example: /housing/550e8400-e29b-41d4-a716-446655440000/7a45d9c2-0000-4000-9000-123456789abc.jpg
        in: query
        name: key
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "200":
          description: Пустой ответ со статусом 200
          schema:
            type: string
        "400":
          description: missing image key / missing housing ID — не передан key или
            id
          schema:
            type: string
        "401":
          description: unauthorized — cookie `token` отсутствует, просрочена или недействительна
          schema:
            type: string
        "500":
          description: failed to delete image — изображение не найдено, объект не
            принадлежит пользователю или удаление не выполнено
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Удалить изображение объекта жилья
      tags:
      - Изображения жилья
    post:
      consumes:
      - multipart/form-data
      description: |-
        Загружает одно или несколько изображений для объекта жилья текущего пользователя.
        ID объекта передаётся в path. Файлы передаются через multipart-поле `file`; поле можно повторять для нескольких файлов.
        Разрешены MIME-типы `image/jpeg`, `image/png`, `image/webp`. Порядок изображений назначается автоматически.
      parameters:
      - description: UUID объекта жилья
        example: 550e8400-e29b-41d4-a716-446655440000
        in: path
        name: id
        required: true
        type: string
      - description: Изображение жилья; можно передать несколько частей с именем `file`
        in: formData
        name: file
        required: true
        type: file
      produces:
      - text/plain
      responses:
        "200":
          description: Пустой ответ со статусом 200
          schema:
            type: string
        "400":
          description: missing housing ID / too large or bad multipart / file required
            / one of files too large / only jpeg/png/webp allowed — ошибка параметров
            или файла
          schema:
            type: string
        "401":
          description: unauthorized — cookie `token` отсутствует, просрочена или недействительна
          schema:
            type: string
        "500":
          description: cannot read file / failed to upload image — файл не прочитан,
            не загружен или объект не принадлежит пользователю
          schema:
            type: string
      security:
      - CookieAuth: []
      summary: Загрузить изображения объекта жилья
      tags:
      - Изображения жилья
...
schemes:
- http
securityDefinitions:
  CookieAuth:
    description: 'Cookie авторизации. Формат заголовка: Cookie: token={jwt}'
    in: header
    name: Cookie
    type: apiKey
swagger: "2.0"